Split the TASTy model and reader into hyperbole.tasty
The TASTy data model (`Tasty.Symbol`, `Tree`, `Definition`, `File`)
and its byte-level reader — now public as `Tasty.parse` — move to a
new base component depending only on digression and gossamer, with the
compiler dependency compile-only: the reader needs no `Context`, no
classpath and no filesystem, so a debugger can feed it TASTy fetched
over a wire without dragging the runtime compiler along. The
`Teletypeable` presentations of symbols and trees stay in
`hyperbole.core` as named toplevel givens, above the table- and
tree-rendering dependencies they need; `hyperbole.stacks` is
unchanged, reaching everything through `core`. (`build.mill` also
gains `vivisection.eval`'s dependency on the new component, for the
commit that follows.)
