+++ lib/hyperbole/src/tasty/hyperbole.Tasty.scala
@@ -33,67 +33,25 @@
 package hyperbole
 
 import anticipation.*
-import dendrology.*
-import denominative.*
 import digression.*
-import escapade.*
-import escritoire.*, columnAttenuation.ignoreAttenuation
 import gossamer.*
-import hieroglyph.*, textMetrics.uniformMetric
-import iridescence.*
 import rudiments.*
-import spectacular.*
 import vacuous.*
 
 import StackTrace.Frame.Kind
 
-// The TASTy data model: what a `.tasty` file says about the definitions it holds, and how
-// symbols and trees within it are presented. Reading a file into this model needs the
-// compiler's unpickler, so that lives in `hyperbole.stacks`; the model itself does not.
+// The TASTy data model: what a `.tasty` file says about the definitions it holds. Reading a
+// file into this model needs the compiler's unpickler classes at runtime — supplied by the
+// caller's classpath, as for `anthology` — but no compiler `Context`, no classpath and no
+// filesystem: `parse` takes raw bytes, so a debugger can feed it TASTy fetched over a wire.
+// Presentation (the `Teletypeable` renderings of symbols and trees) lives in `hyperbole.core`,
+// above this component's dependencies.
 object Tasty:
+  // Reads a `.tasty` file's definitions. Anything unexpected — an unknown version, a truncated
+  // file, an absent section — yields `Unset`: TASTy is only ever a source of extra detail.
+  def parse(data: Data): Optional[File] = TastyFiles(data)
+
   // TastySymbol → Tasty.Symbol
-  object Symbol:
-    import tableStyles.defaultTableStyle
-
-    given teletypeable: (palette: TastyPalette) => Symbol is Teletypeable =
-      symbol =>
-
-        val flags =
-          symbol.flags.stdlib.map: (flag, on) =>
-            if on then e"${Bg(palette.flagOff)}(${Fg(palette.black)}(·${flag}·))"
-            else e"${Fg(palette.flagOff)}($flag)"
-
-          . join(e" ")
-
-        val properties =
-          symbol.properties.stdlib.map: (property, on) =>
-            if on then e"${Bg(palette.propertyOn)}(${Fg(palette.black)}(·${property}·))"
-            else e"${palette.propertyOff}($property)"
-
-          . join(e" ")
-
-        val details =
-          symbol.details.stdlib.map: detail =>
-            detail.absolve match
-              case (key, value: Text) =>
-                key -> e"${Fg(palette.outline)}($value)"
-
-              case (key, items: List[Text] @unchecked) =>
-                key -> e"${Fg(palette.outline)}(${items.join(t", ")})"
-
-          . to(List)
-
-        val name = (t"Name", e"$Bold(${symbol.prefix}${Fg(palette.foreground)}(${symbol.name}))")
-
-        Scaffold[(Text, Teletype)]
-          ( Column(e"$Bold(Property)", textAlign = TextAlignment.Right)(_(0)),
-            Column(e"$Bold(Value)", sizing = columnar.ParagraphOrBreak)(_(1)) )
-
-        . tabulate((name :: (t"Flags", flags) :: (t"Properties", properties) :: details.stdlib).to(List))
-        . grid(120)
-        . render
-        . join(e"\n")
-
   case class Symbol
     ( prefix:     Text,
       name:       Text,
@@ -102,59 +60,6 @@ object Tasty:
       details:    List[(Text, List[Text] | Text)] )
 
   // TastyTree → Tasty.Tree
-  object Tree:
-    import tableStyles.minimalTableStyle
-
-    case class Expansion
-      ( text: Teletype, typeName: Text, param: Optional[Text], expr: Text, source: Teletype )
-
-    private def expand(tree: Tree)(using palette: TastyPalette): List[Expansion] =
-      TreeDiagram.by[Tree](_.nodes)(tree).map: tiles =>
-        node =>
-          val color = (node.term, node.definitional) match
-            case (true, true)   => palette.termDefinition
-            case (false, true)  => palette.typeDefinition
-            case (true, false)  => palette.termReference
-            case (false, false) => palette.typeReference
-
-          val text = e"$color(${node.name})"
-          val tag2: Text = if node.tag == ' ' then "▪".tt else "⟨"+node.tag+"⟩"
-
-          Expansion
-            ( e"${(tiles.stdlib.drop(1).map(treeStyles.defaultTreeStyle.text(_))).to(List).join}$tag2 $text",
-              node.typeName,
-              node.param,
-              node.shortCode,
-              node.source.teletype )
-
-      . stdlib.to(List)
-
-    given teletypeable: (palette: TastyPalette) => Tree is Teletypeable =
-      tastyTree =>
-
-        val expansions = expand(tastyTree)
-
-        val indents =
-          expansions.stdlib.filter(!_.source.nil).map(_.source.plain.keep(_ == ' ').length)
-
-        val crop = indents.min
-
-        Scaffold[Expansion]
-          ( Column(e"TASTy"): node =>
-              val param =
-                node.param.let { param => e"$Italic(${Fg(palette.accented)}($param))" }.or(e"")
-
-              e"${node.text} $param",
-            Column(e"Type"): node =>
-              val name = StackTrace.rewrite(node.typeName.s, false)
-              if node.typeName.nil then e"" else e"${Fg(palette.outline)}(: $Italic(${name}))",
-            Column(e"Source")(_.source.skip(crop)) )
-
-        . tabulate(expansions)
-        . grid(10000)
-        . render
-        . join(e"\n")
-
   case class Tree
     ( tag:          Char,
       typeName:     Text,
