+++ lib/hyperbole/src/tasty/soundness_hyperbole_tasty.scala
@@ -32,5 +32,4 @@
                                                                                                   */
 package soundness
 
-export hyperbole.{Introspect, semantics, syntax, TastyPalette,
-    tastySymbolTeletypeable, tastyTreeTeletypeable}
+export hyperbole.Tasty
